requirements.txt
flask
requests
python-dotenv
gunicorn
SQLAlchemy
psycopg2
